LayoutTests/inspector/debugger/pause-reason.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/pause-reasons.js"></script>
<script>
TestPage.allowUncaughtExceptions = true;
TestPage.needToSanitizeUncaughtExceptionURLs = true;

function test()
{
    WebInspector.debuggerManager.allExceptionsBreakpoint.disabled = false;
    WebInspector.debuggerManager.assertionsBreakpoint.disabled = false;

    for (let script of WebInspector.debuggerManager.dataForTarget(WebInspector.mainTarget).scripts) {
        if (!/pause-reasons\.js$/.test(script.url))
            continue;
        let sourceCodeLocation = script.createSourceCodeLocation(3, 0);
        let breakpoint = new WebInspector.Breakpoint(sourceCodeLocation);
        WebInspector.debuggerManager.addBreakpoint(breakpoint);
        break;
    }

    function sanitizePauseData(data) {
        if (data.description)
            return {description: data.description};
        if (data.breakpointId)
            return {breakpointId: data.breakpointId.substring(data.breakpointId.indexOf("pause-reasons.js"))};
        return data;
    }


    let suite = InspectorTest.createAsyncSuite("Debugger.PauseReason");

    let pauses = 0;

    function addTestCase({name, description, expression, setup}) {
        suite.addTestCase({
            name, description,
            test(resolve, reject) {
                if (setup)
                    setup();

                expression += `;setTimeout(() => { TestPage.dispatchEventToFrontend("AfterTestFunction"); })`;
                InspectorTest.evaluateInPage(expression);

                WebInspector.debuggerManager.singleFireEventListener(WebInspector.DebuggerManager.Event.Paused, (event) => {
                    let targetData = WebInspector.debuggerManager.dataForTarget(WebInspector.debuggerManager.activeCallFrame.target);
                    InspectorTest.log("PAUSE #" + (++pauses));
                    InspectorTest.log("  REASON: " + targetData.pauseReason);
                    if (targetData.pauseData)
                        InspectorTest.log("  DATA: " + JSON.stringify(sanitizePauseData(targetData.pauseData)));
                    else
                        InspectorTest.log("  NO DATA");

                    WebInspector.debuggerManager.resume().catch(reject);
                    InspectorTest.singleFireEventListener("AfterTestFunction", resolve);
                });

                WebInspector.debuggerManager.singleFireEventListener(WebInspector.DebuggerManager.Event.Resumed, function(event) {
                    InspectorTest.log("RESUMED");
                });
            }
        });
    }

    addTestCase({
        name: "Debugger.PauseReason.Exception",
        description: "Pause for exception should have expected pause reason.",
        expression: `setTimeout(triggerException)`,
    });

    addTestCase({
        name: "Debugger.PauseReason.DebuggerStatement",
        description: "Pause for debugger statement should have expected pause reason.",
        expression: "setTimeout(triggerDebuggerStatement)",
    });

    addTestCase({
        name: "Debugger.PauseReason.Assertion",
        description: "Pause for assertion should have expected pause reason.",
        expression: "setTimeout(triggerAssert)",
    });

    addTestCase({
        name: "Debugger.PauseReason.Breakpoint",
        description: "Pause for breakpoint should have expected pause reason.",
        expression: "setTimeout(triggerBreakpoint)",
    });

    addTestCase({
        name: "Debugger.PauseReason.PauseImmediately",
        description: "Pause for Debugger.pause command should have expected pause reason.",
        expression: "setTimeout(() => { 1 + 1 })",
        setup() { WebInspector.debuggerManager.pause(); },
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test that pausing in different ways triggers different pause reasons.</p>
</body>
</html>
